requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
requests
jinja2
python-multipart
